Refactor CRUD to a trait. Added update endpoint
