+++ tests/SolrClientTest.php
@@ -49,26 +49,6 @@ class SolrClientTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test **/
-    public function round_trip()
-    {
-        $solr = new SolrClient('localhost', 8983);
-        $solr->setCore('gettingstarted');
-        $document = new SolrDocument(['id' => 12345, 'title' => 'fish']);
-        $result = $solr->add($document);
-        $this->assertEquals(0, $result['responseHeader']['status']);
-        $solr->commit();
-
-        $doc = $solr->get(12345);
-        $doc->id = 12345;
-
-        $solr->remove(12345);
-        $solr->commit();
-
-        $doc = $solr->get(12345);
-        $this->assertNull($doc);
-    }
-
-    /** @test **/
     public function it_should_commit()
     {
         $solr = new SolrClient('localhost', 8983);
@@ -85,26 +65,5 @@ class SolrClientTest extends PHPUnit_Framework_TestCase
         $result = $solr->optimize();
         $this->assertEquals(0, $result['responseHeader']['status']);
         $this->assertTrue(true);
-    }
-
-    /** @test **/
-    public function it_should_add_and_commit()
-    {
-        $solr = new SolrClient('localhost', 8983, 'gettingstarted');
-        $solr->setAutoCommit(true);
-
-        // add a document, auto commit
-        $solr->add(new SolrDocument(['id' => 1, 'title' => 'test']));
-
-        // get the document right away, make sure it's there
-        $doc = $solr->get(1);
-        $this->assertNotNull($doc);
-
-        // remove the document, auto commit
-        $solr->remove(1);
-
-        // make sure it's gone
-        $doc = $solr->get(1);
-        $this->assertNull($doc);
     }
 }
